error handling improved

+++ src/http/controllers/client.ts
@@ -1,16 +1,17 @@
 import { Request, Response } from 'express';
 import { IClient } from 'domain/client';
 import { CreateClientSchema, UpdateClientSchema } from '../../validation/client';
+import { IDSchema } from '../../validation/id';
 import { ClientService } from '../../services/client';
-import ValidationError from '../../errors/validation';
 import { EHTTP } from '../../enums/http-status-code';
+import ValidationError from '../../errors/validation';
 
 export const getAll = async (_req: Request, res: Response) => {
   try {
     const result = await ClientService.getAll();
     return res.status(EHTTP.StatusOK).json(result);
-  } catch (err) {
-    return res.json(err);
+  } catch (err: any) {
+    return res.status(err.status).json(err);
   }
 };
 
@@ -18,10 +19,13 @@ export const getByID = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
+
     const result = await ClientService.getByID(+id);
     return res.status(EHTTP.StatusOK).json(result);
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
@@ -35,13 +39,16 @@ export const create = async (req: Request, res: Response) => {
     await ClientService.create(client);
     return res.status(EHTTP.StatusCreated).json({ message: 'Created' });
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
 export const update = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
 
     const client = {
       ...req.body,
@@ -54,7 +61,7 @@ export const update = async (req: Request, res: Response) => {
     await ClientService.update(client);
     return res.status(EHTTP.StatusOK).json({ message: 'Updated' });
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
@@ -62,10 +69,13 @@ export const remove = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
+
     await ClientService.remove(+id);
     return res.status(EHTTP.StatusOK).send({ message: 'Removed' });
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
