+++ src/http/controllers/invoice.ts
@@ -1,16 +1,17 @@
 import { Request, Response } from 'express';
 import { IInvoice } from '../../domain/invoice';
 import { CreateInvoiceSchema, UpdateInvoiceSchema } from '../../validation/invoice';
+import { IDSchema } from '../../validation/id';
 import { InvoiceService } from '../../services/invoice';
-import ValidationError from '../../errors/validation';
 import { EHTTP } from '../../enums/http-status-code';
+import ValidationError from '../../errors/validation';
 
 export const getAll = async (_req: Request, res: Response) => {
   try {
     const result = await InvoiceService.getAll();
     return res.status(EHTTP.StatusOK).json(result);
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
@@ -18,10 +19,13 @@ export const getByID = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
+
     const result = await InvoiceService.getByID(+id);
     return res.status(EHTTP.StatusOK).json(result);
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
@@ -39,13 +43,16 @@ export const create = async (req: Request, res: Response) => {
     await InvoiceService.create(client);
     return res.status(EHTTP.StatusCreated).json({ message: 'Created' });
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
 export const update = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
 
     const client = {
       ...req.body,
@@ -60,7 +67,7 @@ export const update = async (req: Request, res: Response) => {
     await InvoiceService.update(client);
     return res.status(EHTTP.StatusOK).json({ message: 'Updated' });
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
 
@@ -68,10 +75,13 @@ export const remove = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
+
     await InvoiceService.remove(+id);
     return res.status(EHTTP.StatusOK).send({ message: 'Removed' });
   } catch (err: any) {
-    return res.status(err.status || EHTTP.StatusInternalServerError).json(err);
+    return res.status(err.status).json(err);
   }
 };
