+++ src/http/controllers/user.ts
@@ -1,16 +1,17 @@
 import { Request, Response } from 'express';
 import { IUser } from 'domain/user';
 import { CreateUserSchema, UpdateUserSchema } from '../../validation/user';
+import { IDSchema } from '../../validation/id';
 import { UserService } from '../../services/user';
-import ValidationError from '../../errors/validation';
 import { EHTTP } from '../../enums/http-status-code';
+import ValidationError from '../../errors/validation';
 
 export const getAll = async (_req: Request, res: Response) => {
   try {
     const result = await UserService.getAll();
     return res.status(EHTTP.StatusOK).json(result);
-  } catch (err) {
-    return res.json(err);
+  } catch (err: any) {
+    return res.status(err.status).json(err);
   }
 };
 
@@ -18,10 +19,13 @@ export const getByID = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
+
     const result = await UserService.getByID(+id);
     return res.status(EHTTP.StatusOK).json(result);
-  } catch (err) {
-    return res.json(err);
+  } catch (err: any) {
+    return res.status(err.status).json(err);
   }
 };
 
@@ -34,14 +38,17 @@ export const create = async (req: Request, res: Response) => {
 
     await UserService.create(user);
     return res.status(EHTTP.StatusCreated).json({ message: 'Created' });
-  } catch (err) {
-    return res.json(err);
+  } catch (err: any) {
+    return res.status(err.status).json(err);
   }
 };
 
 export const update = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
+
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
 
     const user = {
       ...req.body,
@@ -53,8 +60,8 @@ export const update = async (req: Request, res: Response) => {
 
     await UserService.update(user);
     return res.status(EHTTP.StatusOK).json({ message: 'Updated' });
-  } catch (err) {
-    return res.json(err);
+  } catch (err: any) {
+    return res.status(err.status).json(err);
   }
 };
 
@@ -62,9 +69,12 @@ export const remove = async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
 
+    const idValidation = IDSchema.validate(id);
+    if (idValidation.error?.message) throw new ValidationError(idValidation.error?.message);
+
     await UserService.remove(+id);
     return res.status(EHTTP.StatusOK).send({ message: 'Removed' });
-  } catch (err) {
+  } catch (err: any) {
     return res.send(err);
   }
 };
